Add group sender key

+++ signal-service/src/main/java/com/algomeet/signalservice/service/UserDeviceService.java
@@ -11,14 +11,19 @@ import org.springframework.util.StringUtils;
 import com.algomeet.signalservice.dto.UserDeviceRequest;
 import com.algomeet.signalservice.dto.UserDeviceResponse;
 import com.algomeet.signalservice.entity.KyberPreKey;
+import com.algomeet.signalservice.entity.KyberPreKeyId;
+import com.algomeet.signalservice.entity.OneTimePreKey;
 import com.algomeet.signalservice.entity.SignedPreKey;
+import com.algomeet.signalservice.entity.SignedPreKeyId;
 import com.algomeet.signalservice.entity.UserDevice;
 import com.algomeet.signalservice.entity.UserDeviceId;
 import com.algomeet.signalservice.exceptions.RecordNotFoundException;
 import com.algomeet.signalservice.mapper.KyberPreKeyMapper;
+import com.algomeet.signalservice.mapper.OneTimePreKeyMapper;
 import com.algomeet.signalservice.mapper.SignedPreKeyMapper;
 import com.algomeet.signalservice.mapper.UserDeviceMapper;
 import com.algomeet.signalservice.repository.KyberPreKeyRepository;
+import com.algomeet.signalservice.repository.OneTimePreKeyRepository;
 import com.algomeet.signalservice.repository.SignedPreKeyRepository;
 import com.algomeet.signalservice.repository.UserDeviceRepository;
 
@@ -30,8 +35,9 @@ import lombok.RequiredArgsConstructor;
 public class UserDeviceService {
 
 	private final UserDeviceRepository repository;
-	private final SignedPreKeyRepository signedPreKeyServiceRepository;
-	private final KyberPreKeyRepository kyberPreKeyServiceRepository;
+	private final SignedPreKeyRepository signedPreKeyRepository;
+	private final KyberPreKeyRepository kyberPreKeyRepository;
+	private final OneTimePreKeyRepository oneTimePreKeyRepository;
 
 	public UserDeviceResponse createDevice(UUID userKey, UserDeviceRequest request) {
 		// To generate new device ID, get maximum user device ID from table then increment it by 1.
@@ -43,11 +49,15 @@ public class UserDeviceService {
 		
 		//Save prekeys
 		SignedPreKey signedPreKey = SignedPreKeyMapper.toEntity(userKey, deviceId, request.getSignedPreKey());		
-		SignedPreKey savedSignedPreKey= signedPreKeyServiceRepository.save(signedPreKey);
+		SignedPreKey savedSignedPreKey= signedPreKeyRepository.save(signedPreKey);
 		
 		// Save kyber		
 		KyberPreKey kybePreKey = KyberPreKeyMapper.toEntity(userKey, deviceId, request.getKyberPreKey());		
-		KyberPreKey savedKyberPreKey = kyberPreKeyServiceRepository.save(kybePreKey);
+		KyberPreKey savedKyberPreKey = kyberPreKeyRepository.save(kybePreKey);
+		
+		// Save ontime prekeys
+		List<OneTimePreKey> otPreKeys = request.getOneTimePreKeys().stream().map(otp -> OneTimePreKeyMapper.toEntity(userKey, deviceId, otp)).toList();
+		List<OneTimePreKey> savedOtPreKeys = oneTimePreKeyRepository.saveAll(otPreKeys);
 		
 		// Construct response
 		UserDeviceResponse userDeviceResponse = UserDeviceMapper.toResponse(savedDevice);
@@ -55,6 +65,7 @@ public class UserDeviceService {
 
 		userDeviceResponse.setSignedPreKey(SignedPreKeyMapper.toResponse(savedSignedPreKey));
 		userDeviceResponse.setKyberPreKey(KyberPreKeyMapper.toResponse(savedKyberPreKey));
+		userDeviceResponse.setOneTimePreKeys(savedOtPreKeys.stream().map(OneTimePreKeyMapper::toResponse).toList());
 		
 		return userDeviceResponse;
 	}
@@ -82,7 +93,19 @@ public class UserDeviceService {
 	}
 
 	public void deleteDevice(UUID userKey, Integer deviceId) {
-		UserDeviceId id = new UserDeviceId(userKey, deviceId);
-		repository.deleteById(id);
+		UserDeviceId id = new UserDeviceId(userKey, deviceId);		
+		if (repository.findById(new UserDeviceId(userKey, deviceId)).isEmpty()) {
+			throw new RecordNotFoundException("User device ID not found");
+		}
+		
+		SignedPreKeyId signedPreKeyId = new SignedPreKeyId(userKey, deviceId);
+		signedPreKeyRepository.deleteById(signedPreKeyId);
+		
+		KyberPreKeyId kyberPreKeyId = new KyberPreKeyId(userKey, deviceId);
+		kyberPreKeyRepository.deleteById(kyberPreKeyId);
+		
+		oneTimePreKeyRepository.deleteByUserKeyAndDeviceId(userKey, deviceId);
+				
+		repository.deleteById(id);		
 	}
 }
